requirements.txt
absl-py==1.4.0
asgiref==3.6.0
astunparse==1.6.3
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==23.3.0
boto3==1.26.123
botocore==1.29.123
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
# contourpy==1.0.7
contourpy==1.0.6
cryptography==40.0.2
cycler==0.11.0
distlib==0.3.6
Django==4.1.7
django-bootstrap4==23.1
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-environ==0.10.0
django-filter==23.1
django-storages==1.13.2
django-tables2==2.5.3
filelock==3.11.0
flatbuffers==23.5.9
fonttools==4.39.3
gast==0.4.0
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.2
gunicorn==20.1.0
h5py==3.8.0
idna==3.4
imageio==2.27.0
jax==0.4.10
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
keras==2.12.0
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
ml-dtypes==0.1.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.0
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.0
pandas==2.0.1
paramiko==3.1.0
pathspec==0.11.1
Pillow==9.4.0
platformdirs==3.2.0
protobuf==4.23.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyNaCl==1.5.0
pynput==1.7.6
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruff==0.0.254
s3transfer==0.6.0
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
# scipy==1.8
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
storage==0.0.4.3
sympy==1.12
# tensorboard==2.12.3
# tensorboard-data-server==0.7.0
# tensorflow==2.12.0
# tensorflow-estimator==2.12.0
# tensorflow-intel==2.12.0
# tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.1.0
tifffile==2023.3.21
# torch==2.0.1
# torchvision==0.15.2
# torchxrayvision==1.1.1
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
virtualenv==20.21.0
Werkzeug==2.3.4
whitenoise==6.3.0
wrapt==1.14.1
# django-bootstrap-datepicker-plus==5.0.3
# pydicom==2.3.1
# mritopng==2.2
